refactor(recorder): split pure part-event classifiers into FlightRecorder.Classifiers.cs
